Clean up: Merge separate test files to respective sources
