+++ src/math/array_ops_test.ts
@@ -1136,6 +1136,212 @@ const testsClone: MathTests = it => {
   });
 };
 
+const testsTile: MathTests = it => {
+  it('1D (tile)', math => {
+    const t = Array1D.new([1, 2, 3]);
+    const t2 = math.tile(t, [2]);
+
+    expect(t2.shape).toEqual([6]);
+    test_util.expectArraysClose(t2, [1, 2, 3, 1, 2, 3]);
+  });
+
+  it('2D (tile)', math => {
+    const t = Array2D.new([2, 2], [1, 11, 2, 22]);
+    let t2 = math.tile(t, [1, 2]);
+
+    expect(t2.shape).toEqual([2, 4]);
+    test_util.expectArraysClose(t2, [1, 11, 1, 11, 2, 22, 2, 22]);
+
+    t2 = math.tile(t, [2, 1]);
+    expect(t2.shape).toEqual([4, 2]);
+    test_util.expectArraysClose(t2, [1, 11, 2, 22, 1, 11, 2, 22]);
+
+    t2 = math.tile(t, [2, 2]);
+    expect(t2.shape).toEqual([4, 4]);
+    test_util.expectArraysClose(
+        t2, [1, 11, 1, 11, 2, 22, 2, 22, 1, 11, 1, 11, 2, 22, 2, 22]);
+  });
+
+  it('3D (tile)', math => {
+    const t = Array3D.new([2, 2, 2], [1, 2, 3, 4, 5, 6, 7, 8]);
+    const t2 = math.tile(t, [1, 2, 1]);
+
+    expect(t2.shape).toEqual([2, 4, 2]);
+    test_util.expectArraysClose(
+        t2, [1, 2, 3, 4, 1, 2, 3, 4, 5, 6, 7, 8, 5, 6, 7, 8]);
+  });
+
+  it('propagates NaNs', math => {
+    const t = Array1D.new([1, 2, NaN]);
+
+    const t2 = math.tile(t, [2]);
+
+    expect(t2.shape).toEqual([6]);
+    test_util.expectArraysClose(t2, [1, 2, NaN, 1, 2, NaN]);
+  });
+
+  it('1D bool (tile)', math => {
+    const t = Array1D.new([true, false, true], 'bool');
+    const t2 = math.tile(t, [2]);
+
+    expect(t2.shape).toEqual([6]);
+    expect(t2.dtype).toBe('bool');
+    test_util.expectArraysEqual(t2, [1, 0, 1, 1, 0, 1]);
+  });
+
+  it('2D bool (tile)', math => {
+    const t = Array2D.new([2, 2], [true, false, true, true], 'bool');
+    let t2 = math.tile(t, [1, 2]);
+
+    expect(t2.shape).toEqual([2, 4]);
+    expect(t2.dtype).toBe('bool');
+    test_util.expectArraysEqual(t2, [1, 0, 1, 0, 1, 1, 1, 1]);
+
+    t2 = math.tile(t, [2, 1]);
+    expect(t2.shape).toEqual([4, 2]);
+    expect(t2.dtype).toBe('bool');
+    test_util.expectArraysEqual(t2, [1, 0, 1, 1, 1, 0, 1, 1]);
+
+    t2 = math.tile(t, [2, 2]);
+    expect(t2.shape).toEqual([4, 4]);
+    expect(t2.dtype).toBe('bool');
+    test_util.expectArraysEqual(
+        t2, [1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 1, 1, 1]);
+  });
+
+  it('3D bool (tile)', math => {
+    const t = Array3D.new(
+        [2, 2, 2], [true, false, true, false, true, false, true, false],
+        'bool');
+    const t2 = math.tile(t, [1, 2, 1]);
+
+    expect(t2.shape).toEqual([2, 4, 2]);
+    expect(t2.dtype).toBe('bool');
+    test_util.expectArraysEqual(
+        t2, [1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0]);
+  });
+
+  it('bool propagates NaNs', math => {
+    const t = Array1D.new([true, false, NaN] as boolean[], 'bool');
+    const t2 = math.tile(t, [2]);
+
+    expect(t2.shape).toEqual([6]);
+    expect(t2.dtype).toBe('bool');
+    test_util.expectArraysEqual(
+        t2, [1, 0, util.getNaN('bool'), 1, 0, util.getNaN('bool')]);
+  });
+
+  it('1D int32 (tile)', math => {
+    const t = Array1D.new([1, 2, 5], 'int32');
+    const t2 = math.tile(t, [2]);
+
+    expect(t2.shape).toEqual([6]);
+    expect(t2.dtype).toBe('int32');
+    test_util.expectArraysEqual(t2, [1, 2, 5, 1, 2, 5]);
+  });
+
+  it('2D int32 (tile)', math => {
+    const t = Array2D.new([2, 2], [1, 2, 3, 4], 'int32');
+    let t2 = math.tile(t, [1, 2]);
+
+    expect(t2.shape).toEqual([2, 4]);
+    expect(t2.dtype).toBe('int32');
+    test_util.expectArraysEqual(t2, [1, 2, 1, 2, 3, 4, 3, 4]);
+
+    t2 = math.tile(t, [2, 1]);
+    expect(t2.shape).toEqual([4, 2]);
+    expect(t2.dtype).toBe('int32');
+    test_util.expectArraysEqual(t2, [1, 2, 3, 4, 1, 2, 3, 4]);
+
+    t2 = math.tile(t, [2, 2]);
+    expect(t2.shape).toEqual([4, 4]);
+    expect(t2.dtype).toBe('int32');
+    test_util.expectArraysEqual(
+        t2, [1, 2, 1, 2, 3, 4, 3, 4, 1, 2, 1, 2, 3, 4, 3, 4]);
+  });
+
+  it('3D int32 (tile)', math => {
+    const t = Array3D.new([2, 2, 2], [1, 2, 3, 4, 5, 6, 7, 8], 'int32');
+    const t2 = math.tile(t, [1, 2, 1]);
+
+    expect(t2.shape).toEqual([2, 4, 2]);
+    expect(t2.dtype).toBe('int32');
+    test_util.expectArraysEqual(
+        t2, [1, 2, 3, 4, 1, 2, 3, 4, 5, 6, 7, 8, 5, 6, 7, 8]);
+  });
+
+  it('int32 propagates NaNs', math => {
+    const t = Array1D.new([1, 3, NaN], 'int32');
+    const t2 = math.tile(t, [2]);
+
+    expect(t2.shape).toEqual([6]);
+    expect(t2.dtype).toBe('int32');
+    test_util.expectArraysEqual(
+        t2, [1, 3, util.getNaN('int32'), 1, 3, util.getNaN('int32')]);
+  });
+};
+
+const testsGather: MathTests = it => {
+  it('1D (gather)', math => {
+    const t = Array1D.new([1, 2, 3]);
+
+    const t2 = math.gather(t, Array1D.new([0, 2, 0, 1], 'int32'), 0);
+
+    expect(t2.shape).toEqual([4]);
+    test_util.expectArraysClose(t2, [1, 3, 1, 2]);
+  });
+
+  it('2D (gather)', math => {
+    const t = Array2D.new([2, 2], [1, 11, 2, 22]);
+    let t2 = math.gather(t, Array1D.new([1, 0, 0, 1], 'int32'), 0);
+    expect(t2.shape).toEqual([4, 2]);
+    test_util.expectArraysClose(t2, [2, 22, 1, 11, 1, 11, 2, 22]);
+
+    t2 = math.gather(t, Array1D.new([1, 0, 0, 1], 'int32'), 1);
+    expect(t2.shape).toEqual([2, 4]);
+    test_util.expectArraysClose(t2, [11, 1, 1, 11, 22, 2, 2, 22]);
+  });
+
+  it('3D (gather)', math => {
+    const t = Array3D.new([2, 2, 2], [1, 2, 3, 4, 5, 6, 7, 8]);
+
+    const t2 = math.gather(t, Array1D.new([1, 0, 0, 1], 'int32'), 2);
+
+    expect(t2.shape).toEqual([2, 2, 4]);
+    test_util.expectArraysClose(
+        t2, [2, 1, 1, 2, 4, 3, 3, 4, 6, 5, 5, 6, 8, 7, 7, 8]);
+  });
+
+  it('bool (gather)', math => {
+    const t = Array1D.new([true, false, true], 'bool');
+
+    const t2 = math.gather(t, Array1D.new([0, 2, 0, 1], 'int32'), 0);
+
+    expect(t2.shape).toEqual([4]);
+    expect(t2.dtype).toBe('bool');
+    expect(t2.getValues()).toEqual(new Uint8Array([1, 1, 1, 0]));
+  });
+
+  it('int32 (gather)', math => {
+    const t = Array1D.new([1, 2, 5], 'int32');
+
+    const t2 = math.gather(t, Array1D.new([0, 2, 0, 1], 'int32'), 0);
+
+    expect(t2.shape).toEqual([4]);
+    expect(t2.dtype).toBe('int32');
+    expect(t2.getValues()).toEqual(new Int32Array([1, 5, 1, 2]));
+  });
+
+  it('propagates NaNs', math => {
+    const t = Array1D.new([1, 2, NaN]);
+
+    const t2 = math.gather(t, Array1D.new([0, 2, 0, 1], 'int32'), 0);
+
+    expect(t2.shape).toEqual([4]);
+    test_util.expectArraysClose(t2, [1, NaN, 1, 2]);
+  });
+};
+
 const allTests = [
   testsZeros,
   testsOnes,
@@ -1147,6 +1353,8 @@ const allTests = [
   testsRandTruncNormal,
   testsRandUniform,
   testsFromPixels,
+  testsTile,
+  testsGather,
 ];
 
 test_util.describeMathCPU('array_ops', allTests);
